+++ src/components/AddTask/AddTask.tsx
@@ -43,6 +43,20 @@ const AddTaskInput = styled.input`
   color: white;
   background: transparent;
   font-size: 14px;
+
+  
+  &::placeholder {
+    color: red;
+    opacity: 1; /* Firefox */
+  }
+
+  &:-ms-input-placeholder { /* Internet Explorer 10-11 */
+  color: red;
+  }
+
+  &::-ms-input-placeholder { /* Microsoft Edge */
+  color: red;
+  }
 `;
 
 const AddTaskDescription = styled(AddTaskElement)`